requirements.txt
sqlite-utils
markdownify
beautifulsoup4
python-dateutil
PyYAML>=6.0
